requirements-dev.txt
-r ./requirements.txt
pre-commit==2.8.2
django-test-migrations==1.0.0
django-migration-linter==2.4.0
bandit==1.6.2
black==20.8b1
flake8==3.8.4
isort==5.6.4
pytest==6.1.2
pytest-asyncio==0.14.0
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-factoryboy==2.0.3
pytest-lazy-fixture==0.6.3
pytest-testmon==1.0.3
pytest-watch==4.2.0
factory-boy==3.1.0
callee==0.3.1
